+++ packages/presentation-author/src/icons/index.ts
@@ -1,0 +1,4 @@
+export * from './icon-provider.js';
+export * from './prompts.js';
+export * from './types.js';
+export * from './workspace.js';